Resolve "Query parameter parsing and validation"

Closes #77

See merge request hercules/hercules!91

+++ hercules-cassandra-util/src/main/java/ru/kontur/vostok/hercules/cassandra/util/CassandraDefaults.java
@@ -11,7 +11,7 @@ public final class CassandraDefaults {
     public static final int DEFAULT_CASSANDRA_PORT = 9042;
     public static final long DEFAULT_READ_TIMEOUT_MILLIS = 12_000L;
     public static final String DEFAULT_KEYSPACE = "hercules";
-    public static final short DEFAULT_REPLICATION_FACTOR = 3;
+    public static final int DEFAULT_REPLICATION_FACTOR = 3;
 
     // Default pooling options:
